pyproject.toml
[project]
name = "kaflow-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "chromadb>=1.0.20",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-deepseek>=0.1.4",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.7",
    "langsmith>=0.4.26",
    "mcp>=1.0.0",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "ragflow-sdk>=0.20.4",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
]
